==> notebooks/01_Data_Preparation.ipynb <==
### Changing to the main directory

%cd ..

### Importing Necessary Libraries

from utils import DATASET_ID_OR_URL, DATA_DIR, DATA_PATH, PROCESSED_DATA_DIR
from utils import clean_text
import opendatasets as od
import pandas as pd

import os

### Data Ingestion

od.download(DATASET_ID_OR_URL, DATA_DIR)

dataset = pd.read_csv(DATA_PATH)

dataset.sample(n=5, random_state=42)

dataset.drop(columns=["Unnamed: 0"], inplace=True)

### Splitting Dataset into Training, Validation, and Test Sets

from sklearn.model_selection import train_test_split

# Settings for 70 - 15 - 15 split of training, validation, and test datasets
train_df, test_df = train_test_split(dataset, test_size=0.15, random_state=42)
train_df, val_df = train_test_split(train_df, test_size=0.1765, random_state=42)

print(f"Train set shape: {train_df.shape}")
print(f"Validation set shape: {val_df.shape}")
print(f"Test set shape: {test_df.shape}")

train_df.info()

val_df.info()

test_df.info()

There are no missing values in the dataset which indicates its well curated.

### Cleaning the Dataset

train_df['question'] = train_df['question'].apply(clean_text)
val_df['question'] = val_df['question'].apply(clean_text)
test_df['question'] = test_df['question'].apply(clean_text)

train_df['answer'] = train_df['answer'].apply(clean_text)
val_df['answer'] = val_df['answer'].apply(clean_text)
test_df['answer'] = test_df['answer'].apply(clean_text)

### Saving the Dataset

train_df.to_csv(os.path.join(PROCESSED_DATA_DIR, "train.csv"), index=False)
val_df.to_csv(os.path.join(PROCESSED_DATA_DIR, "val.csv"), index=False)
test_df.to_csv(os.path.join(PROCESSED_DATA_DIR, "test.csv"), index=False)
