==> notebooks/02_Data_Tokenization.ipynb <==
### Changing to the main directory

%cd ..

### Importing Necessary libraries

from utils import PROCESSED_DATA_DIR
from utils import tokenized_function

import pandas as pd
import os

from datasets import Dataset

### Loading the cleaned up data

train_df = pd.read_csv(os.path.join(PROCESSED_DATA_DIR, "train.csv"))
val_df = pd.read_csv(os.path.join(PROCESSED_DATA_DIR, "val.csv"))
test_df = pd.read_csv(os.path.join(PROCESSED_DATA_DIR, "test.csv"))

train_df.sample(n=3, random_state=42)

val_df.sample(n=3, random_state=42)

test_df.sample(n=3, random_state=42)

### Tokenization of Datasets

train_data = Dataset.from_pandas(train_df)
val_data = Dataset.from_pandas(val_df)
test_data = Dataset.from_pandas(test_df)

train_tokenized_data = train_data.map(tokenized_function, batched=True)
val_tokenized_data = val_data.map(tokenized_function, batched=True)
test_tokenized_data = test_data.map(tokenized_function, batched=True)

train_tokenized_data = train_tokenized_data.select_columns(["input_ids", "labels"])
val_tokenized_data = val_tokenized_data.select_columns(["input_ids", "labels"])
test_tokenized_data = test_tokenized_data.select_columns(["input_ids", "labels"])

train_tokenized_data

val_tokenized_data

test_tokenized_data

### Saving Tokenized Datasets

train_tokenized_data.save_to_disk(os.path.join(PROCESSED_DATA_DIR, "train_tokenized_data"))
val_tokenized_data.save_to_disk(os.path.join(PROCESSED_DATA_DIR, "val_tokenized_data"))
test_tokenized_data.save_to_disk(os.path.join(PROCESSED_DATA_DIR, "test_tokenized_data"))
