==> notebooks/05_Model_Evaluation_Qualitative_Analysis.ipynb <==
### Changing to the main directory

%cd ..

### Importing Necessary Libraries

from transformers import AutoModelForSeq2SeqLM, AutoTokenizer

from utils import FINAL_MODEL_PATH, DEVICE, BASE_MODEL_PATH, MODEL_ID, PROCESSED_DATA_DIR, SENTENCE_EMBEDDING_MODEL
from utils import full_prompt, clear_gpu_memory, generate_response, cosine_similarity
from sentence_transformers import SentenceTransformer

import pandas as pd

print(full_prompt("What is cuda insight?"))

### Loading Trained Model

trained_model = AutoModelForSeq2SeqLM.from_pretrained(FINAL_MODEL_PATH).to(DEVICE)
tokenizer = AutoTokenizer.from_pretrained(FINAL_MODEL_PATH)
clear_gpu_memory()

generate_response(model=trained_model, tokenizer=tokenizer, prompt="What is cuda insight?")

Lets compare with the original base model

original_model = AutoModelForSeq2SeqLM.from_pretrained(BASE_MODEL_PATH).to(DEVICE)
tokenizer = AutoTokenizer.from_pretrained(MODEL_ID)
clear_gpu_memory()

generate_response(model=original_model, tokenizer=tokenizer, prompt="What is cuda insight?")

Lets pick few examples from test dataset

test_df = pd.read_csv(f"{PROCESSED_DATA_DIR}/test.csv")
test_sample = test_df.sample(n=5, random_state=42)
del test_df

generate_response(model=original_model, tokenizer=tokenizer, prompt="What is cuda insight?")

test_sample['generated_answer_original'] = test_sample['question'].apply(lambda x: generate_response(model=original_model, tokenizer=tokenizer, prompt=full_prompt(x)))
test_sample['generated_answer_trained'] = test_sample['question'].apply(lambda x: generate_response(model=trained_model, tokenizer=tokenizer, prompt=full_prompt(x)))

test_sample

del trained_model
del original_model
clear_gpu_memory()

Lets compare similarity scores between answer and trained answer.

test_sample.drop(columns=['generated_answer_original'], inplace=True)

test_sample.reset_index(drop=True, inplace=True)

# Load the sentence transformer model
sentence_model = SentenceTransformer(SENTENCE_EMBEDDING_MODEL)

embeddings_original = sentence_model.encode(test_sample['answer'].tolist())
embeddings_generated = sentence_model.encode(test_sample['generated_answer_trained'].tolist())

# Calculate cosine similarities and add them to the dataframe
test_sample['cosine_similarity'] = [cosine_similarity(embeddings_original[i], embeddings_generated[i]) for i in range(len(test_sample))]

test_sample

for i in range(5):
    print(f"Question {i+1}: {test_sample['question'][i]}")
    print(f"Actual Answer: {test_sample['answer'][i]}")
    print(f"Predicted Answer: {test_sample['generated_answer_trained'][i]}")
    print(f"Cosine Similarity: {test_sample['cosine_similarity'][i]}")
    print()

### Conclusion

- **Conclusion for Question 1:** The predicted answer captures the essence but misses specific details like "security" and "API stability." The emphasis on fostering and enhancing support aligns with the actual answer.
- **Conclusion for Question 2:** Both answers refer to the simplification of memory management with CUDA's unified memory, though the predicted answer adds the aspect of a virtual address space, which is accurate and provides a bit more technical insight.
- **Conclusion for Question 3:** The predicted answer is circular and doesn't provide a meaningful method for calculating bandwidth. The actual answer gives a clear method for calculation using specific properties.
- **Conclusion for Question 4:** Both answers provide a correct definition of where dependencies are defined in a kit file, with the predicted answer providing an example of a specific dependency.
- **Conclusion for Question 5:** The predicted answer is somewhat redundant and less specific than the actual answer. It states the effect on performance but lacks the how – the details about reducing latency and improving throughput.

Overall, the predicted answers show some alignment with the actual answers but tend to be less specific or miss certain details. The model seems to understand the general context but lacks precision in technical specifics, and in one case, the predicted answer was not useful. The cosine similarity scores reflect the varying degrees of alignment between the predicted and actual answers.

### Suggestions for Improvement

- **Enhance Training Data:** Include more detailed and technical content in the training dataset to improve specificity.
- **Permutation and Combination of Hyper Parameters:** Fine-tune the model's parameters to focus on technical accuracy and detail retention.
- **Use Domain-Specific Models:** Consider training or using a domain-specific model that specializes in technical and programming language.
- **Post-Processing Predictions:** Implement a post-processing step to check for and correct circular or redundant predictions.
- **Evaluation Metrics:** Utilize more robust evaluation metrics that can capture the effectiveness of technical content beyond cosine similarity, like BLEU score for language translation accuracy.
- **User Feedback Loop:** Incorporate user feedback to continuously improve model predictions over time.
- Use base model with more parameter e.g. flan-t5-large
